Fixed some incorrect responses to "TURN".
The PRE-TURN routine was too eager to intercept custom actions. It will
still intercept "TURN" actions on objects that don't have the TURNBIT,
but every other default message is left to V-TURN.

I've reinstated an older default message for turning objects with
something that isn't a tool, but I don't think there's any way to
actually trigger it. I'm just keeping it because it used to be there,
and it's a sensible default message.

This requires some changes to Zork I, II and III, but since they are in
other repositories I'm making separate pull requests for those.